Don't use `ImmutableSortedSet` when `ImmutableSet` will do
Summary:
This is a bug, but it may take some explaining of why it's a bug and what it means (and why it allows me to un-Ignore a seemingly unrelated test).

The goal I'm working towards - and the functionality the newly enabled test is testing - is that using `//foo:bar` in `cquery` should match `//foo:bar` in all configurations. By going into the debugger I found that `targetUniverse.resolveTargetSpecs` is actually doing what I want and returning all `BuildTarget`s that match the spec. Great! Except the test still fails, because later down the line we throw that info away and simply give back targets in one configuration.

I found this was happening when we turned them into `QueryBuildTarget`s. We would place 4 `QueryBuildTarget`s into the builder, but then only get 2 back out. This led me to a relevant passage from the documentation of `ImmutableSortedSet`:

> <p><b>Warning:</b> as with any sorted collection, you are strongly advised not to use a {link
> Comparator} or {link Comparable} type whose comparison behavior is <i>inconsistent with
> equals</i>. That is, {code a.compareTo(b)} or {code comparator.compare(a, b)} should equal zero
> <i>if and only if</i> {code a.equals(b)}. If this advice is not followed, the resulting
> collection will not correctly obey its specification.

Well what do you know, `QueryTarget::compare` gives back `0` if the targets strings are equal, which given the implementation of `QueryBuildTarget::toString` means that `QueryBuildTarget::equals` is not the same as `QueryBuildTarget::compare`.

What makes this silly though is that we don't need this - `buck query` doesn't print it's targets in any given order (and if it _does_ it's handled at a much higher level than in `TargetPatternEvaluator`. There's no reason to try to do this sorting logic. When we remove it `buck cquery` starts behaving as expected (mostly).

Reviewed By: cjhopman

shipit-source-id: 14ae28dc00b1b0111d0b65757e56f2e94375e853