Add `TargetUniverse` interface for query
Summary:
This will be useful for cquery where `--target-universe` is going to be one of the things you can pass via the command line. To make this possible (and coexist with current query) we need to abstract it away via an interface.

This diff takes all the node-parsing logic from current `buck query` and abstracts it away behind an interface. Then `cquery` (which shares the `BuckQueryEnvironment`) can provide an alternative implementation which parses the target graph up front based on the `--target-universe` CLI parameter.

Few notes on this diff:

* Any place that used the parser was abstracted away behind the target universe, to ensure that nothing can go out and randomly start parsing without the target universe being involved.
* For the most part this diff is just a copy paste of the methods involved from `BuckQueryEnvironment` over to the new `TargetUniverse` implementation.
* I tried to avoid teaching the `TargetUniverse` about the concept of `QueryTarget`, so most of the "new code" written here is just shuffling objects between `QueryTarget`s and `BuildTarget`s.

Reviewed By: bobyangyf

shipit-source-id: 0bc92a778cf35e708e85689e4b76da032e88be04