Fix `ownerForFileWithOwnerThatsOutsideTargetUniverseReturnsNothing`
Summary:
In a previous diff I created a helper function to take away the need to keep doing `result.getStdout().trim()` in my assertions. In that helper function I also added a `result.assertSuccess()`, which I assumed would be a noop.

Doing so caused a test to fail. wat.

It just so happens that we were hitting an exception on the _exact_ test whose expected output was `""`, which is why I missed it. When I started using this helper function and asserting success it surfaced this error to me. So this diff fixes the test.

Reviewed By: stepancheg

shipit-source-id: c2f5bd240012a33c9540054fdab18aa518775373