Use `PrecomputedTargetUniverse` even when the `--target-universe` wasn't provided
Summary:
The goal of this is to avoid using the `LegacyQueryUniverse` in `cquery`, even when a `--target-universe` isn't provided.

To accomplish this, we do some not-so-great things to parse the query so we can extract out the targets referenced in said query. We then use those targets as the root of the universe.

Reviewed By: stepancheg

shipit-source-id: 75c4c538236d51488fabe677812dc8b8e976ce10