Use an unsorted Multimap for aggregating multiquery results
Summary:
Right now the multimap is sorted for no discernable reason. Query output is always unsorted. While normally I would be ok with letting sleeping dogs lie, the sorting function is deduplicating targets that it shouldn't and causing problems with cquery tests.

So in this diff, we fix it by not sorting it.

Reviewed By: mzlee

shipit-source-id: 5b1788e018b14ae608890ce99fdd4a1866dd0f37