Add secondary_dex_weight_limit arg to android_binary
Summary:
The buckconfig version of this arg should be sufficient, but adding this just in case. It might be necessary to use this if some apps need to keep the total number of dexes below 100, but other apps break unless this is lowered.

Also using this to fix and re-enable integration tests that relied on setting linear_alloc_hard_limit on the binary.

bypass-lint

Reviewed By: IanChilds

shipit-source-id: 3d000cedd54e9af312e73992be9e6386b264f334