Fix bug with multiple invocations against buckd
Summary:
Previously if different targes (of the same UDR type) were built in subsequent runs, and buckd was used, an error would result.

This is because the KnownUserDefinedRuleTypes instance was not being shared across runs, and it is only populated on parse. If we don't reparse .bzl files, that new instances wouldn't get populated. Make it so that one instance is used properly, now, and make it harder to accidentally reuse a potentially invalid KnownRuleTypesProvider

Reviewed By: bobyangyf

shipit-source-id: d9f5c9369a6b0ff825ac2adb8e50d40f5e052bb9