Consolidate how we provide the `visibility` attribute
Summary:
Right now we have two ways of providing `visibility`/`within_view` when we ask for the raw attributes of a target. The first is what is used by most consumers (eg `buck targets`) and that is the raw value, taken straight from the `BuildFileManifest`. The other is what's used by `query`, which is more correct and takes into account things like the `PACKAGE` files.

This diff consolidates the two around the second, more correct version. Query is no longer a special case, and `buck targets` follows suit. The only downside is that we now need to use `buck audit rules` to get the raw output of the macros.

shipit-source-id: fcf5d978466472a554b105875c99fbf97868c116