Make `UnconfiguredTargetConfiguration.toString()` return the name
Summary:
When you use `TargetConfigurationFactory.create` you can pass the string `builtin//platform:unconfigured` to get back a target configured for the `UnconfiguredTargetConfiguration`. This is the only reference to that string though, and if you do `UnconfiguredTargetConfiguration.toString()` (as is done in the output code of `cquery`) you get the string `UnconfiguredTargetConfiguration`.

Rather than have two different strings that are referring to the same thing (used in two different places) combine the two use cases and make the rule-like form (`builtin//platform:unconfigured`) serve for both.

Reviewed By: stepancheg

shipit-source-id: b010f6813ea37ddcfb566286043d72d9885df796