ParamName type
Summary:
We have a lot of maps by `String`, sometimes these strings are by camel case, and sometimes by snake case which make code harder to read.

Even worse, sometimes we mix camel case and snake case in maps (when we deal with UDR, and `within_view` is another example when we handle names incorrectly).

This type is meant to replace `String` for parameter name.

Probably we should convert everything to snake case, but it's hard to make it safely in one diff.

Reviewed By: mykola-semko

shipit-source-id: f7fb7416a66264740b58fc9a44f5d007c59a4e99