Add ability to define output hash size threshold
Summary:
After Buck creates any Apple bundle (.app, etc), it will hash all contained files. For large bundles, this can take a significant amount of time which becomes prohibitive for *incremental builds*. The computation is also unnecessary as `apple_bundle()` is not an input-based rule, nor is it consumed by any other input-based rule.

In this diff, we add the ability to specify output hash size limits. Previously, input rulekey threshold was used for that purpose. There's a new config section which can be used to specify the thresholds, like so:

  [output_hash_size_limits]

  apple_bundle = 0

NOTE: We preserve the pre-existing behaviour for backwards compatibility reasons.

Reviewed By: cjhopman

shipit-source-id: c12bc593265cb4ced9b60f0fb08823b475242afb