Add support for running external test runner with a tty
Summary:
To enable more interesting user experiences we'd like the external
runner process to have access to a terminal. We can reuse the "command flag"
code used for `buck run` and `buck fix`. Instead of having buck's java code run
the process, we hand it off to the Python wrapper. There we can give the
external full access to the terminal.

Enabling the tty support might not always be what's expected. We guard it
behind a new configuration option `test.external_runner_tty=true/false` that
has to be set explicitly and defaults to `false`.

Reviewed By: bobyangyf

shipit-source-id: d7537a9eadab86dcb7f58af8d4126113387661b9