Add CxxLinkGroupMappingTargetPatternMatcher to support matching build targets
Summary:
We want to support matching build targets in link groups, as this will allow us to define mapping such as:

`("//foo:main_app_binary", "tree", "pattern://bar/...")`

where all targets in `//bar/...` that are in the deps tree of `foo:main_app_binary` will go into a specific link group.

In the previous diff, we made matching generic. Here we add `CxxLinkGroupMappingTargetPatternMatcher` to support pattern matching.

We expect the pattern provided to be a valid build target pattern.

Reviewed By: milend

shipit-source-id: 691ab2acc0e6b39ffce5d60cbcd91e7a9fffd2a6