Fix invalid double-format in Buck logging
Summary:
I was puzzled to see ~1000 of these messages in my Buck log:

```
11:09:48.538][debug][command:3d12d99d-57f0-483a-a35a-06a678b0d5f5][tid:1747][com.facebook.buck.shell.ShellStep] Invalid format string: DEBUG 'Environment: PATH=... HISTTIMEFORMAT=%F %T ...' []
```

Turns out the the log messages we receive in `AppendableLogRecord` are already formatted, but we are attempting to format them a second time. Most of the time this is fine, but if the formatted message happens to include a format specifier, we blow up.

These instances of `AppendableLogRecord` come from exactly one place:

https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/xplat/build_infra/buck_client/src/com/facebook/buck/core/util/log/Logger.java?commit=913c1f207ca7454cdc541d7a00348d71742ef2c9&lines=362-365

We do not call `setParameters()` there, so there will never be parameters to format. Thus the second formatting pass is unnecessary. We know that a normal formatter does nothing because [the docs for `java.util.logging.Formatter`](https://docs.oracle.com/javase/7/docs/api/java/util/logging/Formatter.html#formatMessage(java.util.logging.LogRecord)) explicitly say this:

> If there are no parameters, no formatter is used.

Our mistake is not just unnecessary, but harmful, however; if the already-formatted string contains something that looks like a format specifier, `Formatter#format` will throw. Hence the messages in my console.

I overrode `setParameters()` to enforce the invariant by throwing if a non-empty array of parameters is provided.

I believe D17845685 is the blame revision; it fixed another bug related to format string confusion by pre-formatting the string, but we ended up with a double format.

(See D2765851 for the original introduction of `AppendableLogRecord`.)

Reviewed By: stepancheg

shipit-source-id: 40919d89cb42d6790df92dfda06176a0d83cf8a0