SpecialAttr and use it in query commands
Summary:
`ParamName` now extends `ParamNameOrSpecial`.

`SpecialAttr` class added which also extends `ParamNameOrSpecial`.

`ParamNameOrSpecial` is a key in maps used in implementations or `query` and `targets` commands.

All of this is done to avoid spending brain time thinking which strings are camel case and which strings are snake case, and worrying about accidentally changing one to another.

Reviewed By: mykola-semko

shipit-source-id: 6c4929704bd9b209d2c202e68df68da20fd13456