Extract logic from CxxPreprocessAndCompile.Impl
Summary:
We need to re-use the logic from `CxxPreprocessAndCompile.Impl` to be able to create compilation commands, so let's extract into methods first which will then be moved into a superclass in the next diff.

NOTE: Methods are declared as `protected` as they will move into the superclass in the next diff.

Reviewed By: cjhopman

shipit-source-id: 6e682031ccc7240daf75c8e0c035ddc0cacfda8f