Loosen up parsing logic for detecting --version
Summary:
Due to the way that we intercept --version in the Python wrapper, we have no way of disambiguating flags that take arguments and subcommands. For example:

```
buck --version --oncall foo
```

is parsed as an invocation of the `foo` subcommand, which causes us to bypass the Python wrapper and go into Buck, which gets very confused.

This commit looses up the detection of `is_version` to display the version string even if the wrapper thinks there might be a subcommand, which is generally what we want anyway - `buck --version build` shouldn't do a build, it should just print the version.

Reviewed By: philipjameson

shipit-source-id: 65c2fc4544d996e4f7a8567beb8c46cf7088a50b