Remove unnecessary supplier in AppleCxxPlatforms
Summary:
When I was originally working on D21292065, I had been trying to use a supplier to save constantly checking for the LD version, but in hindsight I realized that I'm recreating a supplier here everytime and using it only once. We don't need to wrap this in a supplier.

Alternatively, if we really do want to cache this, we'd need to do it by the various inputs (`filesystem, xcodeToolFinder, toolSearchPaths, version`). Seems overkill since we already cache the generated platforms and we plan to update this logic anyway.

Reviewed By: milend

shipit-source-id: cdd4c2c0d2776d3785a52d1b3376e2e0631cdb4a