Fix bug with recursive target specs in cquery
Summary:
cquery has its own target spec resolution logic[1] we have to be very careful to make sure we implement it correctly. In this case, we are ignoring the `recursive` field and implementing `//foo/...` as "targets in foo/BUCK" not "targets in all BUCK files in the tree rooted at foo".

This diff fixes it.

[1] I know, terrible idea, but as far as I know we had to since target specs mean something different for cquery than they do for the parser generally.

Reviewed By: stepancheg

shipit-source-id: c928ff1564bb087046651df2d23398fc5b6cfb5e