Update version at which we determine not to set sdk_version
Summary: While the llvm documentation suggests that clang with linker version 520 will prefer the platform args over sdk_version, Xcode 11.3.1 with version 530 does not appear to utilize these platform args correctly. One symptom is that the target SDK is wrong. This is fixed in version 11.4.1 with linker version 556.6. Additionally, with `-fatal_warnings` set, the warning is not treated as an error until 11.4.1 (again, misleading llvm documentation), so we can make the determination to not set it at the higher version where the SDK seems to be doing the right thing.

Reviewed By: milend

shipit-source-id: 7be9206862e21baaebe1ef180c1eaf160e7e25e9