Pass empty scratch directory to SwiftStdlibStep
Summary:
A precondition for `SwiftStdlibStep` is for the temp dir to be empty, otherwise the code does not work correctly.

This meant the step only worked on clean builds (usually CI) but prevented reproduction of bugs locally.

Here, we ensure that the scratch dir passed to `SwiftStdlibStep` is actually empty.

NOTE: We do not want to change the operation of `SwiftStdlibStep`, so that it cleans the temp dir itself, as that's not how the original was setup.

Reviewed By: williamtwilson

shipit-source-id: 8446ca9a1ad916711766ca5ba0135cab8dede1ee