Get rid of TestSelectable
Summary:
`Selectable` is something which:
* can be used in select keys
* can be used in `compatible_with` checks

Production code has only one implementation of `Selectable` which is `ConfigSettingSelectable` which is a set of constraint values.

`TestSelectable` is a dummy `Selectable` implementation used in some tests. This `TestSelectable` object does more harm than good:
* it tests code under assumptions which are not used in production (it handles unspecified platforms differently for example)
* certain code in production Buck assumes that there's only one `Selectable` which is `ConfigSettingSelectable` by downcasting it, for example, "refines" implementation
* it prevents certain changes/optimizations, which we cannot do because we cannot inspect contents of `TestSelectable` objects in production code
* it does not simplify test code much

After this diff `Selectable` interface has only one implementation, which is `ConfigSettingSelectable`.

This diff is a preparation to implement
* `AnySelectable`
* more restrictive `select` keys validation

Reviewed By: bobyangyf

shipit-source-id: c4c7841f8ced3ef7d448be0f713da2364a40ae2d