Update projectV2 generated header map to contain all project targets
Summary:
We generate an Xcode project for a target (TARGET), allowing "focusing" on dependencies of TARGET that we care to have in the workspace. These may include certain tests, etc. for which we include in the generate workspace, but because they are not part of the dependency tree of TARGET, we were not including their headers in the header map generated.

We care to have these headers included as well and in order to make indexing work correctly.

Reviewed By: Haud

shipit-source-id: a7756826f93bca61998c92cb8bf81887e8b7c20d