Don't fail the build if xcode-select fails
Summary:
Throwing a runtime exception here causes any buck invocation to fail on a laptop without Xcode or the Command Line Tools installed with no explanation of how to fix it. A lot of developers will never actually use the command line tools but can't use buck without them.

If the toolchain is needed but not installed and this exception is not thrown, buck will invoke `/usr/bin/clang` during the build which will trigger a graphical prompt to install the Command Line Tools. This is a better user experience than a buck stack trace.

Reviewed By: cute-jumper

shipit-source-id: a0c3bc193aa2b211a4474b4f604ba0331df100af