Exclude build target source paths which are labeled with project_exclude_labels
Summary:
Previously we added support to exclude targets from project generation that contained `project_exclude_labels` labels (D21508010). While this excludes these targets from being generated as Xcode targets, if targets with this label are a depended on by other targets, their output may still be added as part of source or resource for another target.

So here we extend the exclusion behavior to also consider sources/resources to targets, when those targets are being written. This is achieved via:
- extracting out the exclusion logic to ProjectExcludeResolver
- updating the project writer to exclude any sources, resources, etc. for BuildTargetSourcePaths that resolve to targets that are labeled for exclusion

Reviewed By: williamtwilson

shipit-source-id: 892f122db1df27e035c4dcc207523b7ef2ab1b2a