Fix issue when using #headers flavor with Swift
Summary:
There's a special `#headers` flavour (undocumented) that can be used when defining dependencies. Its behaviour is to exclude the linking of artifacts produces by the dependencies (useful to prevent duplicate symbols).

There's an assertion which was added as part of the integration of Swift into `apple_library()` which meant that if a dependency contained Swift sources, its `#headers` flavour would fail an assertion.

In this diff, we remove the incorrect assertion and add tests to verify that everything works as expected, with and without Swift sources in a library which we use via the `#headers` flavour.

Reviewed By: gbarcena

shipit-source-id: 16afbaa3454511ff8b89fbb752ef67d91eddeb5d