Allow top-level read_config in .bzl files for Starlark
Summary:
We allow `read_config` in `.bzl` in Python DSL. This diff allows `read_config` in Starlark.

E. g. this was allowed in Python DSL, but not in Starlark.

```
# utils.bzl

rr = native.read_config("foo", "bar")
```

There are two reasons for that:
* a lot of existing fbsource code does top level `read_config`, this code can be migrated to use functions, but might be slower
* top-level `read_config` unlocks constant propagation optimization in Starlark

This diff depends on the fact that we have not implemented incremental parser cache invalidation on configuration changes (we record accessed config values, but drop them afterwards).

Reviewed By: cjhopman, philipjameson

shipit-source-id: 8e120468cda86449321a620d099e5ed0eba09e23