Add criticalpath package for more sophisticated critical path calculation
Summary:
Today, Buck calculates the critical path using the `CriticalPathEventListener` class. This works, but it's difficult to extend due to its tight coupling with the Buck event bus and the critical path log file format. This commit splits the functionality of `CriticalPathEventListener` into a new package, anchored by `CriticalPathBuilder`, which presents a callback-based interface for calculating the critical path as build rules execute.

In addition to what `CriticalPathEventListener` already does, this commit adds a new feature: in addition to tracking the critical path, each node on the critical path keeps track of its second-longest-running sibling build rule and the delta between that rule's execution time and the critical path time. This is useful when looking at the critical path dataset since it provides a "stickiness" metric for a particular build rule; the higher the delta between the critical path node and the sibling node, the more impact there will be when reducing the execution time of that rule. In this way the delta represents the amount of improvement that is possible for a critical path node before the critical path is dominated by the sibling node.

Future commits will change the `CriticalPathEventListener` to use this class to generate the critical path, as well as record the sibling execution time delta in critical path datasets.

Reviewed By: cjhopman

shipit-source-id: 750084ab864354a268275378466623574c3c0d94