Catch and ignore RejectedExecutionException in WebServerBuckEventListener
Summary:
Letting this exception propagate writes it to the log file, which ends up triggering a pathological case where we write thousands of log messages about stuff that doesn't matter.

For example, P131380258 shows a buck log with 15,000 logged instances of `RejectedExecutionException`, each one including a stack trace.

Catch this and log it only in verbose mode.

Reviewed By: swgillespie

shipit-source-id: dab04a1c50cfe8aca6d619338c1b37b2dbbe940c