Add --rsp-quoting=windows on ClangWindowsCompiler for precompiled header compilation too.
Summary:
When trying to build a precompiled header on Windows with Clang, --rsp-quoting=windows is not getting specifie, resulting in clang not being able to find that precompiled header or other includes that are specified with the Windows path separator \ in the response file.

D15487977 had added the original pre arg file args support, but was not specifying it for the precompiled header path (vitalii-topoliuk added to this review) - I assume because it was not tested.

Reviewed By: akrieger

shipit-source-id: 7969425c49322591b33baaee997fd8079946b860