Check that select keys are unambiguous
Summary:
## Description

Prohibit select expressions like this:

```
select({
  ":linux-arm64": [],
  ":linux-clang": [],
})
```

Before this diff this expression is allowed if *current* platform matches only one of them.

This select is resolved successfully on platforms linux-arm64-gcc or linux-x86_64-clang, but fails on linux-arm64-clang).

Now this select is prohibited unconditionally even on linux-arm64-gcc.

This diff is a part of the effort of improving Buck configuration consistency check.

Users dealing with multiconfiguration projects
* should not have to wait for CI for checking all possible configurations
* some configurations are not even built with diff-check, only hourly, it's even more helpful to fail earlier for these case

The issue will become more serious when we implement OR and AND of config settings.

## The definition

Select key set is unambiguous if when we resolve select for any given platform, we find at most one match.

* a pair of keys is unambiguous if they have the same constraint settings but different constraint values (e. g. linux vs windows)
* a pair of keys is unambiguous if one is more specialized than the other (e. g. linux-arm is more specialized than linux)
* otherwise keys are ambiguous

## Cubic

With this diff, the check that select keys are unambiguous is `O(N * N * M)` where:
* N is the number of select keys
* M is the number of constraints in config setting (e. g. 3 in "windows-clang-x86_64`)

I consider this good enough, because in vast majority of cases select conditions are simple and few.

The algorithm can be made faster for the worst cases, but slower in the most common cases (which is a switch by a single constraint). I suggest we postpone optimizations until we ever need them.

Reviewed By: mykola-semko

shipit-source-id: c9b02387c9237ef12ad4d5b38765266e7b950af8