Use string supplier for `Objects.requireNonNull`
Summary: While profiling cquery I found that ~1/5 of the samples in `PrecomputedTargetUniverse.requireNonNull` were in `String.format`. This is silly because we're not using this string most of the time - theoretically this string should only ever be used once (since we throw an exception when we use it). Use a supplier instead so we can cut down on how much time we spend creating useless strings.

Reviewed By: bobyangyf

shipit-source-id: 0daebf00491a53ff99314cc24b8d8a1e8e77589a