prebuilt_python_library: don't extract by default
Summary:
Change the default behavior of building a `prebuilt_python_library`
to avoid extracting, and instead move this functionality behind a
`extract` flavor.

PyFI's wheels are implemented as a `python_library` which uses
`platform_deps` delegate to the python platform-specific
`prebuilt_python_library` wrapping the actual wheel (backed by a
`remote_file`).  As such, a dep on this `python_library` causing a build
time dep on the backing wheels for all platforms, regardless of
whether they are used or not, which causes unnecessary unpacking
at build time.

Reviewed By: asp2insp

shipit-source-id: 22d53d40d7e1bda6040a5b375583c5a2b1cbc6d3