Make NgContext and Env use disjoint from each other
Summary:
We don't need to pass in both context and environment variables, since the only use of the context is for extracting env vars.
Make it clear in the API to either an env map passed in, or a context, and never both.
Eventually we will completely get rid of the context.
This makes the code a bit cleaner to refactor

Reviewed By: irenewchen

shipit-source-id: 285cd52372c4f50452d346bc5ed728a72945e102