Use actual Multimap for cquery index instead of poor mans approximation
Summary:
Looking at a cquery heap dump showed that the majority of memory usage while creating the indices were `ConcurrentHashMap` nodes. This makes sense because we are using a `ConcurrentHashMap` of `ConcurrentHashMap`s to approximate a `ImmutableSetMultimap` (which we then create later).

This would explain the crazy amount of nodes we are seeing in the profile. They also aren't necessary. Originally I did it this way to avoid having to iterate over all the nodes again (since the list of "all the nodes" is quite large) but looking at it with fresh eyes I realized we are _already_ doing that. After the run of `discoverNewTargetsConcurrently` we then go perform a DepthFirst traversal of the graph to set up the edges. If we are going to traverse all the nodes (serially) anyway, why not just set up the Multimap there? Then we don't have nearly as many `ConcurrentHashMap`s floating around and we can probably avoid a bunch of locking as well.

Reviewed By: bobyangyf

shipit-source-id: cedaed8c5e5815a430d8f48f7052467c3745dc99