Add basic test for XcodeNativeTargetGenerator
Summary:
Unfortunately XcodeNativeTargetGenerator is not well tested. In the next diff, we're going to add a test, but in order to minimize to relevant changes, we can add a basic test to ensure that the test infra is setup.

This will also make it easier to add additional tests in the future.

Reviewed By: milend

shipit-source-id: 0173744a51a5b06fb262a70761f11b0155e36a6d