Use Speculative Parsing consistently throughout the `QueryCommand`s
Summary:
If you squint real hard while looking at the `QueryCommand` implementations you'll notice that it actually creates _two_ parsing contexts - one with speculative parsing enabled and one with it disabled. It then passes these parsing contexts to different places. That's... weird, and probably not what we want. If we want to use speculative parsing then we should use it everywhere, not just in one codepath.

Fix it.

Reviewed By: cjhopman

shipit-source-id: e78a9a874829c8c70a1b54f0f47e5dc7a3fd6a60