Introduce `EvaluatingQueryEnvironment`
Summary:
Right now the methods on `ConfiguredQueryEnvironment` that actually involve parsing queries from the command line are all specific to `ConfiguredQueryEnvironment`. This means if we wanted to make another query implementation (say, `uquery` or `aquery`) we wouldn't be able to reuse any of these methods or the code which invokes them.

This diff is the first step away from that - it takes the relevant methods and puts them on an interface - `EvaluatingQueryEnvironment`

Reviewed By: cjhopman

shipit-source-id: 57f73cc6e9a93990f8281050bc919885b8026062