Avoid unnecessary data copying / computation after creating universe
Summary:
The logs I included have already borne fruit.

Looking through the logs of some particularly expensive queries on CI I found that there was a substantial (~20s) delay between when we print the "Done building universe" log and when we print the "evaluating query" log. There is very little code that happens between those two, but the actual creation of the `PrecomputedTargetUniverse` object falls in the range.

So what's happening here? Well these two constructors are both surprisingly expensive. The constructor for `DirectedAcyclicGraph` includes a precondition for `graph.isAcyclic()` which can take a long time to compute on graphs of 1.5million nodes (approximately what we have on CI). Similarly, the `ImmutableMap.copyOf()` does not simply return an immutable view of a map (since there might be other references floating around) and instead has to iterate over the nodes and create a new map. Again this is expensive when you have ~1.5 million nodes.

So we skip both of those things. This is closer to what `buck query` originally did anyway, since it had to deal with `buildTransitiveClosure` modifying the map out from under it. We simply hold on to mutable references and promise not to mutate them.

Reviewed By: bobyangyf

shipit-source-id: 9556acaf43cce7611343060fa3dbeb9d4bcd8fec