Use `TraversableGraph` rather than `DirectedAcyclicGraph` for query outputs
Summary:
The constructor for `DirectedAcyclicGraph` includes a precondition that checks for cycles. This makes sense given the class name, but can be incredibly expensive for large queries. Right now `(c)query` only uses the `DirectedAcyclicGraph` class for `dot` and `thrift` output, and it's weird that some output formats would check for cycles while others don't. This is also an unintentional penalty to using those output formats - using `--dot` for large queries can give a major perf hit for no discernable reason.

This diff fixes that by simply using a `TraversableGraph` instead. It still implements all the necessary methods for printing but without any of the expensive copying or cycle checking that creating a new `DirectedAcyclicGraph` entails.

Reviewed By: bobyangyf

shipit-source-id: f7e2ce6f0a8d45df35238c63e5a86bf2cad36376