Add basic `UnconfiguredQueryCommand`
Summary:
This isn't really implemented yet, but I want to provide a framework for others to work off of. We can at least get some scaffolding set up and add integration tests and stuff.

Later on I want to add a bunch of tests that exercise problematic use cases so that we can test out any of our potential solutions.

Reviewed By: bobyangyf

shipit-source-id: f093b60f5b10a16c40d3d3a061a0b79d89ecbcc1