Add ability to retrieve UUID of Mach-O files
Summary:
Mach-O files usually have a UUID embedded in them which is set by the linker (usually by hashing). Here, we add a method, `Machos.getUuidIfPresent()`, that allows us to retrieve the UUID.

We will be using it prevent us from re-hashing of large Mach-O executables and re-use the UUID from the linker.

Reviewed By: blackm00n

shipit-source-id: a985c9244979cf469d523bd7876a5ed5932ec4e4