Add rule to compute file hashes, tailored towards the Apple platform
Summary:
# Context

In order to perform incremental actions, we need the ability to determine which input files have changed. In Buck, that's usually done by having a rule which computes the hash of a `SourcePath`.

NOTE: A similar rule already exists (`WriteFileHashCode`) but it's in the Android namespace and does not have explicit handling of Mach-O files.

# Changes

We introduce a new MBR rule, `AppleWriteFileHash` which can output a hash given a `SourcePath`. It has special support for using the UUID from Mach-O files rather than re-hashing the files. For large binaries, the perf wins can be significant.

The Mach-O hashing behaviour of the action is configurable which is expected to be exposed via a configuration option.

Reviewed By: cjhopman

shipit-source-id: afb85d45924c4008abc2df6759adf8a2be82aba4