python: support building source databases
Summary:
This adds a support for building a "source database" for any
Python rule type, via the `#source-db` flavor, which is a JSON
file which contains a) a "sources" map mapping the rule's
modules to their on-disk sources and b) a "dependencies" map
mapping all the rule's transitive dep modules to their on-disk
sources.  When building the source DB, Buck will make sure
sources are built and materialized on disk (but not e.g. resources,
native libraries), so building a `#source-db` should be much
cheaper than e.g. building a Python binary which may depend
on native code.

The idea here is to provide the minimum info needed for external
Python tooling (e.g. type checkers) to run correctly on sources
managed by Buck.

Reviewed By: asp2insp

shipit-source-id: c76d5f9ad47efb8cd7f3502ddbb46b1c1b9f9b79