Reuse bash/rust wrapper UUID if set
Summary:
So we've consistent statistics using the same UUID,
between buck_results, buck_status and buck_builds data
sets.

Make sure subprocesses will not see this uuid, so
we don't get the same UUID if buck invokes buck.

Since the bash/rust wrapper sets this UUID regardless
of whether or not it's present, users of buck
cannot generally use this to set the UUID

Reviewed By: cjhopman

shipit-source-id: f5e748fb024649aa9ae8de5ebd515bc891fd1775