Create an Unordered dot output mode and use it for queries
Summary:
Right now there are two output modes for Dot output - `BFS` and `SORTED`. They both iterate over way too much of the graph, but most importantly the `SORTED` output mode also _sorts every node in the graph before printing it_. Not just the output nodes either, it sorts _every_ node.

On CI we routinely have graphs of hundreds of thousands of nodes, and we just don't care about the sort of the output. All we care about is the final graph, and that comes in just fine unsorted.

Therefore, save ourselves a bunch of work and use `UNORDERED` output for queries.

(NOTE: I deliberatly left the `SORTED` option in for things like the Action Graph)

Reviewed By: cjhopman

shipit-source-id: 0192fc49e0d7a8c8e8bcedbfa03ad737a61606af