Teach `PrebuiltAppleFramework` about cxx flavors
Summary: It doesn't really use apple flavors, but it "supports" them because apple libraries and friends propagate their flavors downwards and if it didn't then parsing would fail. Right now it only supports "Apple" platforms, but that misses all the `cxx` platforms that might exist via the `.buckconfig`. This diff makes it so that those platforms are supported as well.

Reviewed By: cjhopman

shipit-source-id: 4ad575e7204f3e08c49d213d10c75f7b840c2f03