Rename `QueryTarget` to `ConfiguredQueryTarget`
Summary:
This is to pave the way for an eventual `UnconfiguredQueryTarget`, that will represent the things we can run a `uquery` over.

This diff is the result of using the IntelliJ refactor tool to rename `QueryTarget` to `ConfiguredQueryTarget`.

The one notable exception here is that I left `QueryFileTarget` alone as I want to use the same class for `UnconfiguredQueryTarget`. It will be a little weird to have a single class implement both `ConfiguredQueryTarget` and `UnconfiguredQueryTarget` but I think files warrant it.

(I really wish I had made this into an ADT way back when).

Reviewed By: cjhopman

shipit-source-id: 817e55270a15377909be05cf1c267e40ed2561fc