Change to cwd before executing external test runner process
Summary:
The buck wrapper supports executing a command through the argsfile pattern. The
run spec in the file contains the binary to run, the arguments to pass, but
also the cwd to use. For `buck run` the cwd passed by buck is ignored by the
wrapper (but only on !Windows). Judging from the tests in
`test/com/facebook/buck/cli:buck_run` this is the expected behavior.

This diff makes the ignoring of cwd for `buck run` more explicit by changing
the `BuckRunSpec` to contain an `Optional<Path>`. For the run command we don't
set it at all, but for the test command where we expect the cwd to change we
set it to the expected value. In the buck wrapper we then make sure we run from
the expected cwd.

With this change the behavior for running tests with an external runner with or
without the command wrapper is now consistent. Before this change, buck would
invoke the external runner from the directory the user used to invoke buck when
using the argsfile support, but always from the repository root otherwise. Now
the wrapper correctly uses the `cwd` passed through the argsfile, which the
`TestCommand` sets to the repo root.

Reviewed By: bobyangyf

shipit-source-id: 83b6bcc1a2efcd6748e773ff101287a1f1111d4e