Remove Java 9 args on Java 8 and below
Summary: We need this in order to upgrade fbsource to run using Java 11. Unfortunately, Windows hosts always use Java 8 (this is also true for fbcode, which has been using Java 11 for a while). Therefore we need to remove the args that we add for the purposes of building with Java 11 that are not recognised by Java 8.

Reviewed By: mykola-semko

shipit-source-id: 00ac4ff2e9f2237bd95afd7607db03b87951bf30