Add Mach-O symbol binding utilities
Summary:
In this diff, we add a utlity method, `MachoBindInfoReader.parseStronglyBoundSymbols()` that returns the set of strongly bound symbols in a Mach-O executable. This functionality is needed for a future diff that will optimise linking speed.

By design, we're not interested in the actual addresses which require binding but only the set of unique symbols.

Reviewed By: blackm00n

shipit-source-id: 7dd6cbb13fb2e986d4a4999b01053ac86c33f966